requirements.txt
Cython
ruamel.yaml
pandas~=1.3.5
requests~=2.27.1
numpy~=1.21.5
pytz
six
tabulate
jieba~=0.42.1
Pillow~=9.0.1
wordcloud~=1.8.1
snownlp~=0.12.3
matplotlib